+++ packages/zosjobs/__tests__/__unit__/DownloadJobs.unit.test.ts
@@ -297,6 +297,32 @@ describe("DownloadJobs", () => {
                 expect(chunks).toEqual(["test"]);
             });
 
+            it("should allow users to call downloadSpoolContentCommon with correct parameters (with encoding as number)", async () => {
+                let uri: string = "";
+                ZosmfRestClient.getStreamed = jest.fn(async (s: AbstractSession, resource: string, r?: any[], stream?: Writable): Promise<any> => {
+                    uri = resource;
+                    stream?._write("test", "utf-8", jest.fn());
+                });
+                const chunks: any[] = [];
+                const jobFile: IJobFile = JSON.parse(JSON.stringify(jobFiles[0]));
+                const spoolParms: IDownloadSpoolContentParms = {
+                    jobFile: jobFile,
+                    jobid: fakeJobID,
+                    jobname: fakeJobName,
+                    encoding: 1147 as any,
+                    binary: false,
+                    stream: new Writable({write: (chunk) => {
+                        chunks.push(chunk);
+                    }})
+                };
+
+                await DownloadJobs.downloadSpoolContentCommon(fakeSession, spoolParms);
+
+                expect(IO.createDirsSyncFromFilePath).not.toHaveBeenCalled();
+                expect(uri).toContain("?fileEncoding=1147");
+                expect(chunks).toEqual(["test"]);
+            });
+
             it("should allow users to call downloadSpoolContentCommon with correct parameters (with invalid encoding)", async () => {
                 let uri: string = "";
                 ZosmfRestClient.getStreamed = jest.fn(async (s: AbstractSession, resource: string, r?: any[], stream?: Writable): Promise<any> => {